Fix conditional compiling
